Bruker utførSteg for å gjenoppta behandling og flytter sett-på-vent validering ut av StegService
